Customize timer functionality
The "Timer" component was almost rewritten. The base functionality was transmitted from "DigitInput" component to "Timer" and modified.
The "DigitInputBlock" component was created instead of "DigitInput".
In the "Timer" component there are 2 realizations for the main function "handleStartStopClick". Read comments.
Now timer can be set, initiate, paused, resumed and reset.