Customize Alarm functionality
The "Alarm" component was almost rewritten.
Functions for "DigitInputBlock" component were extracted in the separate file "digitInputFunc.js".
Function "addZero" was also extracted in the separate file "addZero.js".
The list of all existing alarms is stored in the map "history" inside the object "alarmHistory" in the file "history.js".
To distinguish between alarms in the map, added module "uuid", which generates unique ID's.
Alarms can be added, enabled/disabled and deleted.
Before rendering, alarms are sorted by their own time and on the screen they are shown from the earliest to the latest (despite the original order when they were added in the history).